Update README.md
To save others time, I've added a short example that uses the config object instead of the connection string, and also shows where to set the value for `enableArithAbort` in the config file.    The purpose is to get folks up and running a bit quicker without having to hunt down some minor problems out of the gate.  

Also, for convenience, it runs from the command line to check the connection.