Fix README to show node 12 no longer supported with 9.x
The README showed that node < 12 support was dropped with 9.x, but based on commit 644480535933aa21d3c0677965656dc21f2949c6 it shows that node <= 12 (including node 12) support was dropped and the release notes associated with the version show the same. Updates README to align with commit message and release notes.